build: land 3 more (sv_game/g_svcmds/cg_consolecmds_mp)
Three more small upstream files. Total POSIX build is now ~88 files.

server/sv_game.cpp: 1 KISAKHACK — (char*)(uintptr_t)sv_gametype->current.integer
  to bridge 32→64 bit DvarValue.integer holding a pointer.
game/g_svcmds.cpp: same pattern for g_banIPs->current.integer.
cgame_mp/cg_consolecmds_mp.cpp: 1 [[maybe_unused]] hex-rays scratch.

Also fixed:
- universal/surfaceflags.h: 2 sites cast 0x80000000 to (int) (done earlier).
- gfx_d3d/r_cinematic.h: d3d9 + binklib includes guarded under _WIN32;
  POSIX gets opaque BINK/BINKTEXTURES/HBINK forward decls so headers
  parse for files that pull r_cinematic.h transitively.

posix_backbone_stubs.cpp additions (~50 entries):
- BG_GetShellshockParms, BG_LoadShellShockDvars, BG_SaveShellShockDvars,
  BG_SetShellShockParmsFromDvars (all stubs; return types tightened to
  match bg_local.h sigs).
- BoxDistSqrdExceeds (returns true), Vec2DistanceSq (real impl),
  MatrixTransformVector43 (real impl).
- CG_ActionSlot{Up,Down}_f, CG_NextWeapon_f, CG_PrevWeapon_f, CG_FxTest,
  CG_FxSetTestPosition, CG_RestartSmokeGrenades,
  CG_VisionSetUpdateTweaksFromFile_{Film,Glow},
  CG_IsScoreboardDisplayed.
- CL_AddReliableCommand.
- Com_GetBspFilename, Com_GetServerDObj, Com_UnloadSoundAliases,
  Com_FindSoundAlias.
- ConcatArgs.
- DObjCreateSkel, DObjGetAllocSkelSize, DObjGetBoneInfo, DObjGetBounds,
  DObjGetModel, DObjGetNumModels, DObjGetTree, DObjIgnoreCollision,
  DObjNumBones, DObjSkelAreBonesUpToDate, DObjSkelExists,
  DObjSkelIsBoneUpToDate.
- G_GetEntityTypeName, G_GetFogOpaqueDistSqrd, G_GetSavePersist,
  G_InitGame, G_ResetEntityParsePoint, G_ShutdownGame.
- NET_IsLocalAddress, Scr_IsValidGameType.
- SV_ClipHandleForEntity, SV_GetMapBaseName, SV_LinkEntity,
  SV_SendServerCommand, SV_SetConfigstring.
- Sys_MillisecondsRaw.
- UI_GetActiveMenu, UI_Popup, UI_SafeTranslateString.
- Globals: g_banIPs, g_dedicated, sv_gametype (dvar* nullptrs).

Now includes <game/game_public.h>, <sound/snd_public.h>,
<universal/com_sndalias.h> so type signatures match upstream.

Removed dups now provided by real files:
- SV_GameCommand, SV_ShutdownGameProgs, TRACK_sv_game → sv_game.cpp

+++ src/cgame/cg_effects_load_obj.cpp
@@ -196,7 +196,9 @@ void __cdecl CG_RegisterImpactEffectsForDir(char *dir, EffectFile *effectFile, c
             if (len >= 0)
             {
                 Hunk_CheckTempMemoryHighClear();
-                buffer = (void *)Hunk_AllocateTempMemoryHigh(len + 1, "CG_RegisterImpactEffects");
+                // KISAKHACK: Hunk_AllocateTempMemoryHigh returns int (offset)
+                // in upstream; bridge through uintptr_t.
+                buffer = (void *)(uintptr_t)Hunk_AllocateTempMemoryHigh(len + 1, "CG_RegisterImpactEffects");
                 FS_Read((unsigned __int8 *)buffer, len, f);
                 FS_FCloseFile(f);
                 *((_BYTE *)buffer + len) = 0;
