+++ src/game/g_svcmds.cpp
@@ -21,7 +21,8 @@ void __cdecl G_ProcessIPBans()
     char *s; // [esp+40Ch] [ebp-4h]
 
     numIPFilters = 0;
-    I_strncpyz(str, (char *)g_banIPs->current.integer, 1024);
+    // KISAKHACK: 32-bit pointer in DvarValue.integer; bridge through uintptr_t.
+    I_strncpyz(str, (char *)(uintptr_t)g_banIPs->current.integer, 1024);
     s = str;
     for (t = str; *t; t = s)
     {
